+++ tests/www/test_logs/dag_for_testing_log_view/scheduled__2017-09-01T00:00:00+00:00/task_for_testing_log_view/1.log
@@ -1,0 +1,1 @@
+Log for testing.
